Write raw image with metadata via Image3D
This seems to slow down writing by a factor of 2, and reading by a factor of 30.

In the old scheme, buffering improved reading speed by a factor of 60 (writing
was already buffered from the start). The new reading code does contain
buffering, but perhaps it hasn't been done properly.

Should write a benchmark.