Enable imageprimaries to read multiple files
This implementation uses a loop over the input files, which pushes the data
found in each file to a vector of all events.

A more elegant solution would be to create chain together the iterators for
multiple input files. Unfortunately the compiler was throwing too many obstacles
in the way.