Try to protect final message from disappearing
Sometimes it fails to appear at the end. Maybe something to do with the
progress bar termination. I haven't managed to understand exactly when,
but it seems that it disappears when the number of files processed
exceeds something between 600 and 700. Very strange.

This hack duplicates the message into progress.finish_with_message()