Change Scattergram::value(zero trues) from 1 to MAX
Returning 1 gave maximal importance to LORs in regions where no trues
were detected, which emphasized the scatters.

Returning MAX means that these LORs, which are pretty much guaranteed to
be scatters, are suppressed very heavily.

This completely eliminates the huge accumulation of noise in the corners
of the 2D toy reconstructions, when scatter correction in r is used.