Stop writing mlem test images to shared directory
If two different users ran these tests on the same machine, the second
one would not have permission to write the output files, because the
target directory would already exist and be owned by the first user.

We don't want to use clash-avoiding generated filenames for this,
because the whole point of writing these out is so that the human can
inspect the results visually, so the human should be able to find the
results easily, without having to search for obscurely-located files.

To we just stick them in a subdirectory of the current directory.