Prototype parallel Scattergram filling
This drastically *slows down* the process, because `rayon`
creates *lots* of jobs, far more than there are threads. Each job
requires a new empty Scattergram to be constructed and eventually summed
into the accumulator. This is very expensive. Need to find a way to
limit the number of jobs that `rayon` creates.