Restrict number of rayon jobs with with_min_len()
This makes `rayon` improve rather than degrade performance WRT
the single-threaded version.

Some measurements:

860 million LORs on frontend1petalo

Single-threaded: ~180s
30 threads     :   15s
60 threads     :   11s

Reading the LORs from file (~70s cold; ~45s warm) is now the slowest
part of the LOR-loading chain. So we should probably move on. But I'm
pretty sure that we can do quite a bit better here.

One thing that *does* need to be done right now: the user currently has
to pick the number of threads to use during this step, on the CLI. We
need a decent default.