Prototype rayon subjob limitation in mlem
Rayon dynamically spawns subjobs. It's quite eager about crating lots of
small jobs. Each such job requires creating a new empty accumulator
value, and then combining that accumulator with other accumulators
resulting from other subjobs.

In mlem, the accumulator is an `Image`. These are expensive to construct
and combine, so lots of little jobs lead to lots of overhead.

So here we try to limit Rayon to spawning only one job per thread, by
demanding that no job should be smaller than

  total number of LORs / number of threads

It would be better if we could say precisely what we mean: "one job per
thread", but the current solution already seems to improve performance
appreciably.

We use an unsafe static, in order to avoid passing the number of threads
from main, all the way up the stack to where the work is done.