Place upper limit on rayon subjob size
Here are some rough timings for reconstruction of 860 million LORs on
frontend1petalo. The times given are per single iteration.

- without limits

  + 25 threads ~ 6 s
  + 40 threads ~13 s
  + 70 threads ~30 s

- with min only

  + 25 threads 4.9 s
  + 40 threads 3.7 s
  + 70 threads 2.5 s

- with min and max

  + 25 threads 4.7 s
  + 40 threads 3.0 s
  + 70 threads 2.2 s

And here are the timings for 100 million LORs. This time the timings are
for the whole reconstruction.

- without limits

  + 25 threads    295s
  + 40 threads    660s
  + 70 threads > 1100s

- with min only

  + 25 threads  58s 59s 59s
  + 40 threads  44s 44s 41s
  + 70 threads  37s 38s 37s

- with min and max

  + 25 threads  58s 58s 57s
  + 40 threads  40s 44s 46s 43s 43s
  + 70 threads  36s 35s 35s

So introducing the upper limit *might* improve things *slightly* for the
bigger reconstruction, but it's not clear at all.

But the big gain came when the minimum size was introduced. Without the
lower limit, adding more threads (beyond some fairly small optimum
number) degrades performance; with the lower limit, adding more threads
improves performance.