WIP: unify flow of alternative backproj size data
The prototype implementation of independent forward and backward
projection sizes, passed (optional)

+ FOV dimensions
+ sensitivity matrix row buffer

via independent paths. These needed to by synchronized and kept coherent
in different places in the code, which is very fragile. There was a bug
somewhere, which I haven't

This is an attempt to keep these bits of information together in a
single `Option` throughout, in order to simplify the task of keeping
them coherent.